move tests to shared experiments folder
