+++ src/shared/model/experiments/index.ts
@@ -1,0 +1,4 @@
+import { ABTest } from '@guardian/ab-core';
+import { abTestTest } from './ab-test-test';
+
+export const tests: ABTest[] = [abTestTest];
